accounts: start admindash and lock routes

+++ fastapi-traveltwo/routers/venues.py
@@ -6,13 +6,17 @@ from authenticator import authenticator
 
 router  = APIRouter()
 
+
 # Admin
 @router.post("/api/categories/", response_model=CategoryOut)
 def create_category(
     category: CategoryIn,
     repo: CategoryRepository = Depends(),
+    account_data: dict = Depends(authenticator.get_current_account_data),
 ):
-    return repo.create(category)
+    if account_data['is_admin'] == True:
+        return repo.create(category)
+
 
 # User
 @router.get("/api/categories/", response_model=list[CategoryOut])
@@ -21,6 +25,7 @@ def get_all_categories(
 ):
     return repo.get_all_categories()
 
+
 # User
 @router.post("/api/venues/", response_model=VenueOut)
 def create_venues(
@@ -28,10 +33,12 @@ def create_venues(
     request: Request,
     response: Response,
     repo: VenueRepository = Depends(),
-    # account_data: dict = Depends(authenticator.get_current_account_data),
+    account_data: dict = Depends(authenticator.get_current_account_data),
 ):
+    added_by = account_data['id']
     approved = False
-    return repo.create(venue, approved)
+    return repo.create(venue, added_by, approved)
+
 
 # Admin and Maybe User
 @router.get("/api/venues/{venue_id}", response_model=Optional[VenueOut])
@@ -45,11 +52,13 @@ def get_one_venue(
         response.status_code = 404
     return venue
 
+# Admin to approve venues
 @router.get("/api/venues/unapproved/", response_model=list[VenueOut])
 def get_unapproved_venues(
     repo: VenueRepository = Depends(),
 ):
     return repo.get_unapproved()
+
 
 # Admin
 @router.get("/api/venues/", response_model=list[VenueOut])
@@ -57,6 +66,7 @@ def get_all(
     repo: VenueRepository = Depends(),
 ):
     return repo.get_all()
+
 
 # User
 @router.get("/api/venues/{state}/{city}", response_model=list[VenueCompleteOut])
@@ -67,19 +77,25 @@ def get_all_approved(
 ):
     return repo.get_all_complete(state, city)
 
-# Admin
+
+# Admin to approve venues
 @router.put("/api/venues/{venue_id}", response_model=Union[VenueOut, Error])
 def update_venue(
     venue_id: int,
     venue: VenueIn,
     repo: VenueRepository = Depends(),
+    account_data: dict = Depends(authenticator.get_current_account_data),
 ) -> Union[Error, VenueOut]:
-    return repo.update(venue_id, venue)
+    if account_data['is_admin'] == True:
+        return repo.update(venue_id, venue)
 
-# Admin
+
+# Admin to delete venues
 @router.delete("/api/venues/{venue_id}", response_model=bool)
 def delete_venue(
     venue_id: int,
     repo: VenueRepository = Depends(),
+    account_data: dict = Depends(authenticator.get_current_account_data),
 ) -> bool:
-    return repo.delete(venue_id)
+    if account_data['is_admin'] == True:
+        return repo.delete(venue_id)
